refactor: Move FetherApp methods into methods folder
