+++ packages/fether-electron/src/main/app/methods/index.js
@@ -1,0 +1,60 @@
+// Copyright 2015-2018 Parity Technologies (UK) Ltd.
+// This file is part of Parity.
+//
+// SPDX-License-Identifier: BSD-3-Clause
+
+import setupAppListeners from './setupAppListeners';
+import createWindow from './createWindow';
+import updateProgress from './updateProgress';
+import createPositioner from './createPositioner';
+import setupRequestListeners from './setupRequestListeners';
+import createTray from './createTray';
+import loadTray from './loadTray';
+import showTrayBalloon from './showTrayBalloon';
+import setupDebug from './setupDebug';
+import setupSecurity from './setupSecurity';
+import setupLogger from './setupLogger';
+import setupParityEthereum from './setupParityEthereum';
+import setupGlobals from './setupGlobals';
+import setupMenu from './setupMenu';
+import getScreenResolution from './getScreenResolution';
+import calculateWindowPosition from './calculateWindowPosition';
+import onTrayClick from './onTrayClick';
+import fixWindowPosition from './fixWindowPosition';
+import showWindow from './showWindow';
+import moveWindowUp from './moveWindowUp';
+import processSaveWindowPosition from './processSaveWindowPosition';
+import setupWindowListeners from './setupWindowListeners';
+import setupWin32Listeners from './setupWin32Listeners';
+import hideWindow from './hideWindow';
+import windowClear from './windowClear';
+import onWindowClose from './onWindowClose';
+
+export {
+  setupAppListeners,
+  createWindow,
+  updateProgress,
+  createPositioner,
+  setupRequestListeners,
+  createTray,
+  loadTray,
+  showTrayBalloon,
+  setupDebug,
+  setupSecurity,
+  setupLogger,
+  setupParityEthereum,
+  setupGlobals,
+  setupMenu,
+  getScreenResolution,
+  calculateWindowPosition,
+  onTrayClick,
+  fixWindowPosition,
+  showWindow,
+  moveWindowUp,
+  processSaveWindowPosition,
+  setupWindowListeners,
+  setupWin32Listeners,
+  hideWindow,
+  windowClear,
+  onWindowClose
+};
