+++ packages/fether-electron/src/main/app/methods/onWindowClose.js
@@ -1,0 +1,11 @@
+// Copyright 2015-2018 Parity Technologies (UK) Ltd.
+// This file is part of Parity.
+//
+// SPDX-License-Identifier: BSD-3-Clause
+
+function onWindowClose () {
+  this.processSaveWindowPosition();
+  this.windowClear();
+}
+
+export default onWindowClose;
